Fixed FileSize() for files larger than 2GB
People who used stat_ before might want to watch out, it now uses a long for size instead of an int. But then again, why the hell are you even using stat_?

Also, I've deleted a weird '.filesystem.bmx' in the same directory which
made life a lot more painful than it should have been. It doesn't seem to
be required by anything and looks exactly the same as filesystem.bmx, so
it's probably the by-product of a fat-fingered file copy or something.